Changing the file architecture
