+++ edit/readme_edit_submission.php
@@ -79,6 +79,7 @@ if (isset($_POST['readme_save_button'])) {
     // else {
     //   echo "Sde table name didn't update.";
     // }
+    unset($_SESSION["edit-readme"]);
   }
