+++ export/export_metadata.php
@@ -1,5 +1,5 @@
 <?php
-include ('navbar.php');
+include ('../navbar.php');
 ?>
 
 <style>
