+++ export/expxml_dict.php
@@ -1,7 +1,6 @@
-
 <?php
-include('sqltoxml.php'); //code for coverting a table into xml
-include 'config.php'; //connect to database
+include('../sqltoxml.php'); //code for coverting a table into xml
+include '../config.php'; //connect to database
 $sde_underscore = $_GET['sde_underscore'];
 //the header is used to download the xml file with the name passed in the header
 header('content-type: text/xml');
